backend/pyproject.toml
[project]
name = "tsc-backend"
version = "0.1.0"
description = "TSC Backend - Multi-agent software engineering agency"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "deepagents>=0.4.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-openrouter>=0.1.0",
    "python-dotenv>=1.0.0",
    "langgraph-checkpoint-sqlite>=3.0.0,<4.0.0",
    "aiosqlite>=0.19.0,<1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools", "state"]

[dependency-groups]
dev = [
    "msgpack>=1.1.2",
    "pyinstaller>=6.19.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "setuptools>=82.0.1",
]
